+++ lib/omq/readable.rb
@@ -12,7 +12,7 @@ module OMQ
     # @raise [IO::TimeoutError] if read_timeout exceeded
     #
     def receive
-      with_timeout(@options.read_timeout) { @engine.dequeue_recv }
+      Reactor.run { with_timeout(@options.read_timeout) { @engine.dequeue_recv } }
     end
 
     # Waits until the socket is readable.