+++ lib/omq/routing/radio.rb
@@ -121,7 +121,7 @@ module OMQ
 
 
       def start_group_listener(conn)
-        @tasks << Reactor.spawn_pump(annotation: "recv pump") do
+        @tasks << @engine.spawn_pump_task(annotation: "group listener") do
           loop do
             frame = conn.read_frame
             next unless frame.command?