+++ lib/omq/routing/scatter.rb
@@ -68,7 +68,8 @@ module OMQ
       # @param conn [Connection]
       #
       def start_reaper(conn)
-        @tasks << Reactor.spawn_pump(annotation: "reaper") do
+        return if conn.is_a?(Transport::Inproc::DirectPipe)
+        @tasks << @engine.spawn_pump_task(annotation: "reaper") do
           conn.receive_message
         rescue *CONNECTION_LOST
           @engine.connection_lost(conn)